Pluggable packet module proposal
Now, packet module will be pluggable, as it will try do dynamically
import an implementation specified in config or provided as an argument.
In that way PTF will become independent regarding Scapy, since one can
develop his/her own packet manipulation framework and then create an
implementation of packet module for it. By default it will be Scapy.
